requirements.txt
django-user-accounts==1.3.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
django==1.9.8
pinax-theme-bootstrap==7.10.1
djangorestframework==3.5.4